+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "crates/net/downloaders",
     "crates/primitives",
     "crates/revm",
+    "crates/revm/revm-primitives",
     "crates/rlp",
     "crates/rlp/rlp-derive",
     "crates/rpc/ipc",
