+++ crates/revm/revm-primitives/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "reth-revm-primitives"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/paradigmxyz/reth"
+description = "core reth specific revm utilities"
+
+[dependencies]
+# reth 
+reth-primitives = { path = "../../primitives" }
+
+revm = { version = "3.0.0" }
